+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v3.0.9
-- **[fix]** - `ListAvatar` improvements
-- **[fix]** - `FileViewer` improvements
+## v3.0.10
+- **[fix]** - `ListAvatar` thumbnail improvements 
+- **[fix]** - `FileViewer` thumbnail improvements